[Navigation] Rename navigate to navigation
