+++ app/src/main/java/org/dicio/dicio_android/skills/navigation/NavigationInfo.java
@@ -1,6 +1,6 @@
-package org.dicio.dicio_android.skills.navigate;
+package org.dicio.dicio_android.skills.navigation;
 
-import static org.dicio.dicio_android.SectionsGenerated.navigate;
+import static org.dicio.dicio_android.SectionsGenerated.navigation;
 
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
@@ -13,24 +13,24 @@ import org.dicio.skill.SkillInfo;
 import org.dicio.skill.chain.ChainSkill;
 import org.dicio.skill.standard.StandardRecognizer;
 
-public class NavigateInfo extends SkillInfo {
-    public NavigateInfo() {
-        super("navigate", R.string.skill_name_navigate,
-                R.string.skill_sentence_example_navigate, R.drawable.ic_navigate_white,
+public class NavigationInfo extends SkillInfo {
+    public NavigationInfo() {
+        super("navigation", R.string.skill_name_navigation,
+                R.string.skill_sentence_example_navigation, R.drawable.ic_navigate_white,
                 false);
     }
 
     @Override
     public boolean isAvailable(final SkillContext context) {
-        return Sections.isSectionAvailable(navigate);
+        return Sections.isSectionAvailable(navigation);
     }
 
     @Override
     public Skill build(final SkillContext context) {
         return new ChainSkill.Builder()
-                .recognize(new StandardRecognizer(Sections.getSection(navigate)))
-                .process(new NavigateProcessor())
-                .output(new NavigateOutput());
+                .recognize(new StandardRecognizer(Sections.getSection(navigation)))
+                .process(new NavigationProcessor())
+                .output(new NavigationOutput());
     }
 
     @Nullable
