+++ app/src/main/res/values/strings.xml
@@ -89,10 +89,10 @@
     <string name="skill_sentence_example_open">Open NewPipe</string>
     <string name="skill_name_calculator">Calculator</string>
     <string name="skill_sentence_example_calculator">What is five times four minus a million?</string>
-    <string name="skill_name_navigate">Navigate</string>
-    <string name="skill_sentence_example_navigate">Navigate to Vancouver international airport</string>
-    <string name="skill_navigate_specify_where">Specify where you want to navigate to</string>
-    <string name="skill_navigate_navigating_to">Navigating to %1$s</string>
+    <string name="skill_name_navigation">Navigation</string>
+    <string name="skill_sentence_example_navigation">Navigate to Vancouver international airport</string>
+    <string name="skill_navigation_specify_where">Specify where you want to navigate to</string>
+    <string name="skill_navigation_navigating_to">Navigating to %1$s</string>
     <string name="skill_name_telephone">Telephone</string>
     <string name="skill_sentence_example_telephone">Call Tom</string>
     <string name="skill_name_timer">Timer</string>
