+++ app/src/main/sentences/en/navigation.dslf
@@ -1,3 +1,3 @@
-navigate: high
+navigation: high
 (navigate me?)|(take|bring me)|((give me?)? direction<s?>) to|toward|in? the? .where.;
 where s|is .where.;
