+++ app/src/main/sentences/it/navigation.dslf
@@ -1,3 +1,3 @@
-navigate: high
+navigation: high
 naviga|vai|porta<m|c>i|guida<m|c>i a<l|llo|lla|i|gli|lle?>|verso|in|ne<l|llo|lla|i|gli|lle>? .where.;
 dov<e?> e? .where.;
